[Refactor] Home page의 클라이언트 컴포넌트를 분리
