+++ src/app/[locale]/page.tsx
@@ -1,20 +1,25 @@
-'use client';
+import { Metadata } from 'next';
+import HomeClient from '@/components/HomeClient';
 
-import Lottie from 'react-lottie-player';
-import lottieData from '../../../public/images/loading.json';
-
-const Home = () => {
-  return (
-    <main className="relative w-full h-[calc(100vh-156px)] lg:h-[calc(92vh-60px)] overflow-hidden flex items-center justify-center">
-      <div className="text-center">
-        <h1 className="text-2xl">프로젝트 명</h1>
-        <p className="">프로젝트 간단 소개</p>
-      </div>
-      <div className="absolute bottom-50% lg:right-0 w-[300px] h-[300px] z-0 lg:rotate-90">
-        <Lottie loop animationData={lottieData} play className="lg:rotate-180" />
-      </div>
-    </main>
-  );
+export const metadata: Metadata = {
+  title: '임시 제목',
+  description: '서울 열린데이터광장 공공데이터 활용 창업 경진대회',
+  openGraph: {
+    title: '임시 제목',
+    description:
+      '서울의 야경명소, 전통시장, 유적지를 다국어로 소개하는 여행자 맞춤형 안내 웹사이트입니다.',
+    images: [
+      {
+        url: '/images/seoul_logo.png',
+        width: 1200,
+        height: 630,
+        alt: '홈페이지 오픈그래프 이미지',
+      },
+    ],
+    type: 'website',
+  },
 };
 
-export default Home;
+export default function HomePage() {
+  return <HomeClient />;
+}
